+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.2.0-rc.2
 
+* Force multi-line comments to the next line (#241).
 * Allow splitting after argument name (#368).
 * Parsing a statement fails if there is unconsumed input (#372).
 * Allow splitting on `as` and `is` expressions (#384).